update to the BlastExpression approach
Commands are no longer types/records, with a specific `BlastExpression` type which bounds the output record type. Another small change is to opt for `Seq[String]` instead of `String` everywhere; this makes possible to accomodate commands such as `git add` or logical options which actually group a set of them from the command line perspective.